requirements.txt
python-binance
TA-Lib
numpy
backtrader
backtrader[plotting]
flask
pyfolio
pymc3
theano
seaborn
websocket-client
pandas
mataplotlib



#run the command
#pip install -r  requirements.txt